refactor(quantization): refactor ci by using shell script and change quantization code for format check
