+++ .gitignore
@@ -1,7 +1,6 @@
 *log*/
 *.jpg
 *.png
-*.txt
 
 # compilation and distribution
 __pycache__
