+++ official/quantization/test.py
@@ -16,6 +16,7 @@ Note:
 import argparse
 import time
 
+# pylint: disable=import-error
 import models
 
 import megengine as mge
@@ -113,7 +114,7 @@ def worker(world_size, args):
         valid_dataset,
         sampler=valid_sampler,
         transform=T.Compose(
-            [T.Resize(256), T.CenterCrop(224), T.Normalize(mean=128), T.ToMode("CHW"),]
+            [T.Resize(256), T.CenterCrop(224), T.Normalize(mean=128), T.ToMode("CHW")]
         ),
         num_workers=args.workers,
     )
