+++ requires-style.txt
@@ -1,0 +1,3 @@
+flake8==3.7.9
+isort==4.3.21
+pylint==2.5.2
